Protobuf: Use {telemetry, amd_uav_flight_update}.proto
To be able to use more recent versions of Protobuf, it is needed to regenerate the protocols.
If the version between the pregenerated ones and the installed Protobuf versions differs, you normally get an build error from a macro.
Using a more recent version of Protobuf might be interesting in situations, where things are breaking e.g. when using more modern compilers.

Signed-off-by: Thomas Karl Pietrowski <thopiekar@gmail.com>