Don't override VERSION if already set
When building from git without any tag, you will get a checksum from the method setting VERSION.
However, CMake will complain that the checksum isn't a valid version, thus we want to set the version manually.

This commit will allow this and skip the git magic, if VERSION is set.

Signed-off-by: Thomas Karl Pietrowski <thopiekar@gmail.com>